move container under structures directory
