+++ src/lib/structures/pieces/command.piece.ts
@@ -6,7 +6,7 @@ import {
 } from 'discord.js';
 import { Guards } from '../../types/guards.enum';
 import Piece, { PieceOptions } from './piece';
-import Container from '../../container';
+import Container from '../container';
 
 export interface CommandOptions extends PieceOptions, ChatInputApplicationCommandData {
   guards?: (keyof typeof Guards)[];
